refactor: update main test
