+++ test/main/factories/user/external-sign-in/external-sing-in-validation-factory.spec.ts
@@ -1,7 +1,7 @@
-import { ValidationComposite, RequiredFieldsValidation } from '@/adapters/util/validations';
-import { makeExternalSignInValidationFactory } from '@/main/factories/external-sign-in';
+import { ValidationComposite, RequiredFieldsValidation } from '@/adapters/controllers/util/validations';
+import { makeExternalSignInValidationFactory } from '@/main/factories/user';
 
-jest.mock('@/adapters/util/validations/validation-composite');
+jest.mock('@/adapters/controllers/util/validations/validation-composite');
 
 describe('ExternalSignIn validation Factory', () => {
   it('Should call ValidationComposite with all validations', () => {
