+++ test/main/factories/user/sign-up/sing-up-validation-factory.spec.ts
@@ -1,7 +1,7 @@
-import { makeSignUpValidationFactory } from '@/main/factories/sign-up';
-import { ValidationComposite, RequiredFieldsValidation, CompareFieldsValidation } from '@/adapters/util/validations';
+import { makeSignUpValidationFactory } from '@/main/factories/user';
+import { ValidationComposite, RequiredFieldsValidation, CompareFieldsValidation } from '@/adapters/controllers/util/validations';
 
-jest.mock('@/adapters/util/validations/validation-composite');
+jest.mock('@/adapters/controllers/util/validations/validation-composite');
 
 describe('SignUp validation Factory', () => {
   it('Should call ValidationComposite with all validations', () => {
